+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [[NEXT]](https://github.com/iExecBlockchainComputing/iexec-commons-poco/releases/tag/vNEXT) 2023
 
+### New Features
+- Add `isEligibleToContributeAndFinalize` method to `TaskDescription`. (#53)
 ### Quality
 - Do not run tests in `itest` task to avoid executing them twice. (#47)
 - Upgrade to Gradle 8.2.1 with up-to-date plugins. (#49)